requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sentence-transformers>=2.7.0
faiss-cpu>=1.7.4
pypdf>=4.2.0
numpy>=1.26.0
openai>=1.30.0
python-multipart>=0.0.9
tqdm>=4.66.0

